requirements.txt
antlr4-python3-runtime==4.9.3
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.2
hydra-core==1.3.2
importlib-resources==5.12.0
kiwisolver==1.4.4
matplotlib==3.7.1
numpy==1.24.2
omegaconf==2.3.0
opencv-python-headless==4.5.5.64
packaging==23.0
Pillow==9.4.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
python-dateutil==2.8.2
PyYAML==6.0
qimage2ndarray==1.10.0
six==1.16.0
zipp==3.15.0
